requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
black==24.3.0
coverage==7.2.3
flake8>=7.0.0
matplotlib==3.7.1
gin-config==0.5.0
pytorch-ignite==0.5.0.post2
torch==2.6.0+cu118
lightning==2.4.0
torchmetrics==1.0.3
lightgbm==4.6.0
xgboost==2.1.0
imbalanced-learn==0.12.3
catboost==1.2.5
numpy==1.24.3
pandas==2.2.2
polars==1.9.0
pyarrow==14.0.1
pytest==7.3.1
scikit-learn==1.5.0
tensorboard==2.12.2
tqdm==4.66.3
einops==0.6.1
hydra-core==1.3
optuna==4.0.0
optuna-integration==4.0.0
wandb==0.17.3
recipies==1.0
scikit-optimize-fix==0.9.1
hydra-submitit-launcher==1.2.0
pytest-runner==6.0.1
shap==0.49.1
numba==0.60.0
